examples/requirements.txt
pandas==1.2.1
PyYAML==6.0
